Task 3 & 4 - Get your salary & Give yourself a rise

+++ src/com/company/Car.java
@@ -7,7 +7,7 @@ public class Car {
     final String type;
     public String color;
     public Integer maxSpeed;
-    public Integer price;
+    public Double price;
 
 
     Car(String manufacturer, String model, String type){
